Tighter implementation with Cent and Dollar struct
