+++ Cargo.toml
@@ -1,11 +1,12 @@
 [workspace]
-members = ["app", "db", "tui"]
+members = ["app", "db", "shared", "tui"]
 
 resolver = "2"
 
 [workspace.dependencies]
 db = { path = "db" }
 app = { path = "app" }
+shared = { path = "shared" }
 chrono = "0.4.41"
 diesel = { version = "2.2.12", default-features = false, features = [
     "returning_clauses_for_sqlite_3_35",
