+++ app/src/fetcher/mod.rs
@@ -8,6 +8,6 @@ mod tx_view;
 pub use activity_view::*;
 pub use chart_view::*;
 pub use search_view::*;
-pub use summary_models::*;
+pub(crate) use summary_models::*;
 pub use summary_view::*;
 pub use tx_view::*;
