+++ app/src/modifier/mod.rs
@@ -1,9 +1,11 @@
 mod delete_tx;
+mod new_activity;
 mod new_method;
 mod new_tx;
 mod shared;
 
 pub(crate) use delete_tx::*;
+pub(crate) use new_activity::*;
 pub(crate) use new_method::*;
 pub(crate) use new_tx::*;
 pub use shared::*;
