+++ db/Cargo.toml
@@ -4,6 +4,7 @@ version = "0.1.0"
 edition = "2024"
 
 [dependencies]
+shared.workspace = true
 libsqlite3-sys = { version = "0.35.0", features = ["bundled"] }
 diesel.workspace = true
 chrono.workspace = true
